requirements.txt
google-genai
Flask
python-docx
markdown
python-dotenv